access()の戻り値を受け取るsize_t iをintに

+++ include/pipex.h
@@ -6,7 +6,7 @@
 /*   By: hioikawa <hioikawa@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/01 01:04:46 by hioikawa          #+#    #+#             */
-/*   Updated: 2022/09/02 16:34:05 by hioikawa         ###   ########.fr       */
+/*   Updated: 2022/09/02 16:46:36 by hioikawa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,7 +108,7 @@ char	**split_virtual_argv_from_real_argv(\
 			t_pipex *pipex, char *command_from_argv);
 char	**join_file_and_directry_name_to_get_absolute_path(\
 			t_pipex *pipex, char **list_of_directry, char *command_name);
-size_t	get_index_accessible_path(\
+int	get_index_accessible_path(\
 			t_pipex *pipex, char **list_absolute_path_of_command);
 char	**switch_first_argv_to_absolute_path(\
 			t_pipex *pipex, t_v_argv *v, size_t index);
